requirements.txt
fastapi==0.98.0
langchain==0.0.209
openai==0.27.7
pydantic==1.10.9
python-dotenv==1.0.0
Requests==2.31.0
uvicorn==0.22.0
pandas
reportlab
# googlemaps
pdfplumber
google-cloud-vision
pyspellchecker